feat: Created AI handler
